+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - raid content type is now only disabled for Party frames
 - the sound selection under Settings is now scrollable, containing 20 items before having to scroll
 - synchronize min gap with max frame dimensions
+- added new setting to show fractions of cast durations
+- added support to use DandersFrames if present
 
 ## Version 1.0.0-alpha4